Consolidate claim extraction
