+++ graphrag/index/operations/extract_covariates/__init__.py
@@ -2,10 +2,3 @@
 # Licensed under the MIT License
 
 """The Indexing Engine text extract claims package root."""
-
-from graphrag.index.operations.extract_covariates.extract_covariates import (
-    ExtractClaimsStrategyType,
-    extract_covariates,
-)
-
-__all__ = ["ExtractClaimsStrategyType", "extract_covariates"]
